clean code
# message auto-generated for no-merge-commit merge:
merge clean_code_last into master

clean code 遗留修改

Created-by: msprofDeveloper

Match-id-9990e8f7d11c5a53fc502c3467c99fd8736ae601

+++ collector/dvvp/adx/stub/adx_prof_api.cpp
@@ -43,7 +43,7 @@ int DoAdxIdeCreatePacket(CmdClassT type, IdeString value, uint32_t valueLen, Ide
     }
 
     *buf = static_cast<IdeMemHandle>(sendBuf);
-    *bufLen = static_cast<int>(sizeof(struct tlv_req)) + valueLen;
+    *bufLen = static_cast<int>(sizeof(struct tlv_req)) + static_cast<int>(valueLen);
 
     return IDE_DAEMON_OK;
 }
